merge from main

+++ backend/src/dap-an/dap-an.service.ts
@@ -4,7 +4,7 @@ import { CreateDapAnDto, DapAnDto } from './dto/create-dap-an.dto';
 import { UpdateDapAnDto } from './dto/update-dap-an.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { DapAn } from './entities/dap-an.entity';
-import { EntityManager, Repository } from 'typeorm';
+import { EntityManager, In, Repository } from 'typeorm';
 import { CauHoiService } from 'src/cau-hoi/cau-hoi.service';
 import { LoaiCauHoi } from 'src/common/enum/loaicauhoi.enum';
 import { CreateMotDapAn } from './dto/create-mot-dap-an.dto';
@@ -48,30 +48,31 @@ export class DapAnService {
       }
   }
 
-  findAll() {
-    return `This action returns all dapAn`;
-  }
-
-  findOne(id: number) {
-    return `This action returns a #${id} dapAn`;
-  }
   async timNhieuDapAnTheoIdCauHoi (id:number){
     return await this.dapAnRepo.find({
       where: {idCauHoi: id}
     })
 
   }
+  async xoaTatCaDapAnTheoIdCauHoi(idCauHoi:number){
+    const mangDapAn = await this.timNhieuDapAnTheoIdCauHoi(idCauHoi)
+    const manIdDapAN = mangDapAn.map(da => da.id)
+    try{
+      return await this.dapAnRepo.delete({ idCauHoi: In(manIdDapAN) });
+    }catch(err){
+      throw new InternalServerErrorException('Lỗi xóa đáp án theo id câu hỏi')
+    }
+  }
 
-async capNhatMotDapAn(id: number, dto: UpdateDapAnDto, manager?: EntityManager) {
-  const repo = manager ? manager.getRepository(DapAn) : this.dapAnRepo;
+async capNhatMotDapAn(id: number, dto: UpdateDapAnDto) {
 
   // preload trả về Promise<DapAn | null>
-  const entity = await repo.preload({ id, ...dto });
+  const entity = await this.dapAnRepo.preload({ id, ...dto });
   if (!entity) throw new NotFoundException('Không tìm thấy đáp án');
 
- ;
-   try{
-       return await repo.save(entity)
+  try{
+      await this.dapAnRepo.save(entity)
+      return{messages: "oke"}
      }catch(err){
       throw new InternalServerErrorException('Lỗi sửa một đáp án')
      }
@@ -79,12 +80,15 @@ async capNhatMotDapAn(id: number, dto: UpdateDapAnDto, manager?: EntityManager) 
 
 
   async xoaMotDApAnTheoIdDapAn(id: number) {
-     const cauHoi = await this.cauHoiService.timCauHoiTheoId(id)
-     const dapAn = await this.dapAnRepo.findOne({where: {id}})
+
+     const dapAn = await this.dapAnRepo.findOne({where: {id}});
+     if(!dapAn) throw new NotFoundException('Không tìm thấy đáp án');
+     const cauHoi = await this.cauHoiService.timCauHoiTheoId(dapAn.idCauHoi);
      if(dapAn?.dapAnDung === true && cauHoi.cauHoi.loaiCauHoi === LoaiCauHoi.MotDung)
        throw new BadRequestException('Không thể xóa đáp án đúng của câu hỏi một đúng cần phải thêm câu hỏi đúng thay thế vào')
      try{
-      return await this.dapAnRepo.delete(id);
+       await this.dapAnRepo.delete(id);
+       return{messages: "oke"} 
      }catch(err){
       throw new InternalServerErrorException('Lỗi xóa một đáp án')
      }
